Delete Ray operator tooling
Stage 5 of the Ray retirement plan (#4453). Ray operator scripts in
scripts/ray/ are superseded by scripts/iris/; the cluster Dockerfile
and its CI flow are no longer needed now that Iris handles cluster
lifecycle.

Deletes
- scripts/ray/{cluster,dev_tpu,cleanup_workers}.py and scripts/ray/README.md
- scripts/debug/inspect_data.py (Ray-only data-inspection CLI)
- docker/marin/Dockerfile.cluster (based on rayproject/ray)
- .github/workflows/marin-cleanup-tpus.yaml (invoked scripts/ray/cleanup_workers.py)
- Marin cluster image job in .github/workflows/docker-images.yaml
- Makefile cluster_docker* targets (built the above Dockerfile)

Helpers moved
- _normalize_cluster_region and _validate_data_region lifted from the
  deleted scripts/debug/inspect_data.py into their sole importer,
  tests/test_inspect_data_region.py.

Not deleted this stage
- lib/marin/src/marin/cluster/config.py still imported by
  lib/marin/src/marin/run/ray_run.py, which is queued for stage 3h.
- lib/marin/src/marin/cluster/gcp.py still imported by scripts/gcp-ssh
  and scripts/iris/dev_tpu.py for the non-Ray gcloud helpers
  (get_project_id, find_tpu_by_ip, find_vm_by_ip, ssh_to_vm, ssh_to_tpu).
  Both files need a follow-up after stage 3h to split Ray-specific
  helpers from the generic GCP utilities.

Verification
- rg 'import ray' lib/ scripts/ returns only lib/fray/** hits.
- uv sync (default) succeeds; uv sync --package marin --extra tpu
  resolves 597 packages (libtpu wheel platform mismatch on macOS is
  unrelated to this change).
- ./infra/pre-commit.py --all-files --fix passes.
- Migrated test_inspect_data_region.py: 7 passed in 0.27s.
