+++ .github/workflows/docker-images.yaml
@@ -16,138 +16,6 @@ permissions:
   packages: write
 
 jobs:
-  marin-cluster-images:
-    # Run on: manual trigger OR weekly schedule (02:00 UTC on Sundays)
-    if: |
-      github.event_name == 'workflow_dispatch' ||
-      (github.event_name == 'schedule' && github.event.schedule == '0 2 * * 0')
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: 'actions/checkout@v5'
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v4
-
-      - name: Set tags
-        id: set-tags
-        run: |
-          echo "DATE_TAG=`date -u +"%Y%m%d"`" >> "$GITHUB_OUTPUT"
-          echo "HASH_TAG=`git rev-parse --short HEAD`" >> "$GITHUB_OUTPUT"
-
-      - name: Authenticate to Google Cloud
-        id: auth
-        uses: google-github-actions/auth@v2
-        with:
-          token_format: access_token
-          credentials_json: ${{ secrets.GCP_ARTIFACT_REPOSITORY_SA_KEY }}
-          service_account: github-artifact-repository@hai-gcp-models.iam.gserviceaccount.com
-
-      - name: Login to Google Artifact Registry in europe-west4
-        uses: docker/login-action@v4
-        with:
-          registry: europe-west4-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Login to Google Artifact Registry in us-central1
-        uses: docker/login-action@v4
-        with:
-          registry: us-central1-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Login to Google Artifact Registry in us-central2
-        uses: docker/login-action@v4
-        with:
-          registry: us-central2-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Login to Google Artifact Registry in us-east1
-        uses: docker/login-action@v4
-        with:
-          registry: us-east1-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Login to Google Artifact Registry in us-east5
-        uses: docker/login-action@v4
-        with:
-          registry: us-east5-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Login to Google Artifact Registry in us-west4
-        uses: docker/login-action@v4
-        with:
-          registry: us-west4-docker.pkg.dev
-          username: oauth2accesstoken
-          password: ${{ steps.auth.outputs.access_token }}
-
-      - name: Build Marin cluster image
-        run: |
-          docker buildx build --file docker/marin/Dockerfile.cluster \
-            --provenance=false \
-            --load \
-            --tag marin_cluster:local \
-            .
-
-      - name: Verify Marin cluster image
-        # TODO (rav): should we just run tests inside the container?
-        run: |
-          echo "Verifying critical imports work correctly..."
-          docker run --rm marin_cluster:local python -c "
-          import sys
-          print(f'Python: {sys.version}')
-
-          import numpy
-          print(f'numpy: {numpy.__version__}')
-
-          import pandas
-          print(f'pandas: {pandas.__version__}')
-
-          import ray
-          print(f'ray: {ray.__version__}')
-
-          import torch
-          print(f'torch: {torch.__version__}')
-
-          print('All critical imports verified successfully!')
-          "
-
-      - name: Push Marin cluster image
-        run: |
-          REGISTRIES="europe-west4 us-central1 us-central2 us-east1 us-east5 us-west4"
-          TAGS="latest ${{ steps.set-tags.outputs.DATE_TAG }} ${{ steps.set-tags.outputs.HASH_TAG }}"
-
-          for registry in $REGISTRIES; do
-            for tag in $TAGS; do
-              docker tag marin_cluster:local ${registry}-docker.pkg.dev/hai-gcp-models/marin/marin_cluster:${tag}
-              docker push ${registry}-docker.pkg.dev/hai-gcp-models/marin/marin_cluster:${tag}
-            done
-          done
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: "0.7.20"
-          enable-cache: true
-
-      - name: Update Docker image tag in cluster config script
-        run: sed -i -E 's/LATEST = "[a-z0-9]+"/LATEST = "${{ steps.set-tags.outputs.HASH_TAG }}"/' lib/marin/src/marin/cluster/config.py
-
-      - name: Generate cluster configs
-        run: uv run scripts/ray/cluster.py update-configs
-
-      - name: Create pull request
-        uses: peter-evans/create-pull-request@v6
-        with:
-          commit-message: Update Docker image tag
-          title: Update Docker image tag
-          branch: actions/update-docker-image-tag
-          body: Auto-generated from GitHub Actions.
-
   # Iris Images - Worker, controller, and task images pushed to GHCR
   # Uses a single multi-stage Dockerfile (lib/iris/Dockerfile) with --target.
   # Controller and worker share all layers up to Python deps; the worker adds
