--- .github/workflows/marin-cleanup-tpus.yaml
@@ -1,45 +1,0 @@
-name: Marin - Cleanup Preempted TPUs
-
-on:
-  schedule:
-    - cron: '0 */6 * * *' # Every 6 hours
-
-  workflow_dispatch:
-
-jobs:
-  cleanup-tpus:
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    concurrency:
-      group: cleanup-tpus
-      cancel-in-progress: true
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v5
-
-      - name: Set up Python 3.12
-        uses: actions/setup-python@v6
-        with:
-          python-version: "3.12"
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v7
-        with:
-          enable-cache: true
-
-      - name: Install dependencies
-        run: uv sync --all-packages --no-default-groups
-
-      - name: Authenticate to Google Cloud
-        uses: google-github-actions/auth@v2
-        with:
-          credentials_json: ${{ secrets.GCP_SA_KEY }}
-
-      - name: Set up Google Cloud SDK
-        uses: google-github-actions/setup-gcloud@v2
-        with:
-          project_id: ${{ secrets.GCP_PROJECT_ID }}
-
-      - name: Cleanup preempted and terminated TPUs
-        run: uv run python scripts/ray/cleanup_workers.py --threshold=0
