--- scripts/debug/inspect_data.py
@@ -1,787 +1,0 @@
-#!/usr/bin/env python3
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
-
-"""Inspect training data at a given step.
-
-Given a WandB run and a step number, dump the decoded training examples for
-that step's batch to JSONL. Requires a Ray cluster (--cluster).
-
-Usage:
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --step 100 --cluster us-central2
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --step 100 --cluster us-central2 --output step_100.jsonl
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --step 100 --cluster us-central2 --summary
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --steps 0,100,500 --cluster us-central2 --summary
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --step 100 --cluster us-central2 --truncate 200
-  uv run scripts/debug/inspect_data.py --wandb-run marin-community/marin/<run-id> \
-      --step 100 --cluster us-central2 --tui
-
-  # Also accepts full WandB URLs:
-  uv run scripts/debug/inspect_data.py \
-      --wandb-run https://wandb.ai/marin-community/marin/runs/<run-id> \
-      --step 100 --cluster us-central2 --summary
-"""
-
-import json
-import logging
-import os
-import re
-import shlex
-import sys
-from collections import Counter
-
-import click
-import haliax as hax
-import jax.random as jrandom
-
-from levanter.data.mixture import MixtureDataset, rescale_mixture_schedule_for_batch_schedule
-from levanter.utils.thread_utils import blocking_wait
-
-logger = logging.getLogger(__name__)
-
-
-def _fetch_wandb_config(wandb_run: str) -> dict:
-    """Fetch the raw config dict from a WandB run.
-
-    This must run on a machine with WandB credentials (i.e. the client, not the
-    cluster). Returns a minimal dict that can be serialized to JSON and passed
-    to the cluster via ``--wandb-config-json``.
-
-    Args:
-        wandb_run: WandB run path like "entity/project/run_id" or a full URL.
-    """
-    import wandb
-
-    # Parse URL if provided
-    url_match = re.search(r"wandb\.ai/([^/]+)/([^/]+)/runs/([^/?]+)", wandb_run)
-    if url_match:
-        wandb_run = f"{url_match.group(1)}/{url_match.group(2)}/{url_match.group(3)}"
-
-    api = wandb.Api()
-    run = api.run(wandb_run)
-    c = run.config
-
-    data_cfg = c.get("data", {})
-    trainer_cfg = c.get("trainer", {})
-
-    return {
-        "components": {name: comp["cache_dir"] for name, comp in data_cfg.get("components", {}).items()},
-        "train_weights": data_cfg.get("train_weights"),
-        "tokenizer": data_cfg.get("tokenizer", "gpt2"),
-        "mixture_block_size": data_cfg.get("mixture_block_size", 2048),
-        "stop_strategy": data_cfg.get("stop_strategy", "restart"),
-        "shuffle": data_cfg.get("shuffle", True),
-        "enforce_eos": data_cfg.get("enforce_eos", True),
-        "block_cross_document_attention": data_cfg.get("block_cross_document_attention", True),
-        "train_batch_size": trainer_cfg.get("train_batch_size", 128),
-        "num_train_steps": trainer_cfg.get("num_train_steps", 1),
-        "seed": trainer_cfg.get("seed", 0),
-        "data_seed": c.get("data_seed"),
-        "train_seq_len": c.get("train_seq_len"),
-        "run_id": trainer_cfg.get("id", wandb_run.split("/")[-1]),
-    }
-
-
-def _config_from_wandb_dict(d: dict):
-    """Build a TrainLmConfig from a serialized WandB config dict.
-
-    Returns (TrainLmConfig, run_id).
-    """
-    from levanter.data.text import DatasetComponent, LMMixtureDatasetConfig
-    from levanter.main.train_lm import TrainLmConfig
-    from levanter.tracker.wandb import WandbConfig
-    from levanter.trainer import TrainerConfig
-
-    components = {name: DatasetComponent(cache_dir=cache_dir) for name, cache_dir in d["components"].items()}
-
-    data = LMMixtureDatasetConfig(
-        components=components,
-        train_weights=d["train_weights"],
-        tokenizer=d["tokenizer"],
-        mixture_block_size=d["mixture_block_size"],
-        stop_strategy=d["stop_strategy"],
-        shuffle=d["shuffle"],
-        enforce_eos=d["enforce_eos"],
-        block_cross_document_attention=d["block_cross_document_attention"],
-        auto_build_caches=False,
-    )
-
-    trainer = TrainerConfig(
-        tracker=WandbConfig(entity="marin-community", project="marin"),
-        train_batch_size=d["train_batch_size"],
-        num_train_steps=d["num_train_steps"],
-        seed=d["seed"],
-    )
-
-    config = TrainLmConfig(
-        data=data,
-        trainer=trainer,
-        train_seq_len=d["train_seq_len"],
-        data_seed=d.get("data_seed"),
-    )
-
-    logger.info(
-        "Reconstructed config: %d components, batch_size=%d, seed=%d",
-        len(components),
-        trainer.train_batch_size,
-        trainer.seed,
-    )
-    return config, d["run_id"]
-
-
-def _compute_outlier_scores(text: str) -> dict:
-    """Compute data quality scores for a single example.
-
-    Returns a dict of score names to values. Higher values indicate more
-    anomalous text.
-    """
-    if not text:
-        return {"unique_word_ratio": 0.0, "non_ascii_ratio": 0.0, "digit_ratio": 0.0, "repetition_ratio": 0.0}
-
-    words = text.split()
-    unique_word_ratio = len(set(words)) / len(words) if words else 0.0
-
-    non_ascii_ratio = sum(1 for c in text if ord(c) > 127) / len(text)
-    digit_ratio = sum(1 for c in text if c.isdigit()) / len(text)
-
-    # Check for repetitive 50-char chunks
-    chunk_size = 50
-    chunks = [text[i : i + chunk_size] for i in range(0, max(1, len(text) - chunk_size), chunk_size)]
-    repetition_ratio = 1.0 - len(set(chunks)) / len(chunks) if len(chunks) > 1 else 0.0
-
-    return {
-        "unique_word_ratio": round(unique_word_ratio, 3),
-        "non_ascii_ratio": round(non_ascii_ratio, 3),
-        "digit_ratio": round(digit_ratio, 3),
-        "repetition_ratio": round(repetition_ratio, 3),
-    }
-
-
-def _is_outlier(scores: dict) -> str | None:
-    """Return a human-readable reason if the scores indicate an outlier, else None."""
-    reasons = []
-    if scores["unique_word_ratio"] < 0.3:
-        reasons.append(f"unique_word_ratio={scores['unique_word_ratio']}")
-    if scores["non_ascii_ratio"] > 0.4:
-        reasons.append(f"non_ascii={scores['non_ascii_ratio']:.0%}")
-    if scores["digit_ratio"] > 0.5:
-        reasons.append(f"digits={scores['digit_ratio']:.0%}")
-    if scores["repetition_ratio"] > 0.3:
-        reasons.append(f"repetition={scores['repetition_ratio']:.0%}")
-    return ", ".join(reasons) if reasons else None
-
-
-def _submit_to_cluster(
-    cluster: str,
-    step: int | None,
-    output: str | None,
-    steps: str | None = None,
-    truncate: int | None = None,
-    summary: bool = False,
-    wandb_config_json: str | None = None,
-):
-    import asyncio
-
-    from ray.job_submission import JobSubmissionClient
-
-    from fray.v1.cluster.ray import DashboardConfig, ray_dashboard
-
-    config = _resolve_cluster_config(cluster)
-
-    parts = ["python", "scripts/debug/inspect_data.py"]
-    if wandb_config_json:
-        parts.extend(["--wandb-config-json", wandb_config_json])
-    if steps:
-        parts.extend(["--steps", steps])
-    else:
-        parts.extend(["--step", str(step)])
-    if output:
-        parts.extend(["--output", output])
-    if truncate is not None:
-        parts.extend(["--truncate", str(truncate)])
-    if summary:
-        parts.append("--summary")
-    entrypoint = shlex.join(parts)
-
-    runtime_env = _cluster_runtime_env()
-
-    async def _run():
-        client = JobSubmissionClient("http://127.0.0.1:8265")
-        submission_id = client.submit_job(entrypoint=entrypoint, runtime_env=runtime_env)
-        click.echo(f"Submitted Ray job: {submission_id}", err=True)
-        async for lines in client.tail_job_logs(submission_id):
-            print(lines, end="")
-
-    with ray_dashboard(DashboardConfig.from_cluster(config)):
-        asyncio.run(_run())
-
-
-def _build_dataset(train_config):
-    """Build the MixtureDataset from a resolved train config."""
-    seed = train_config.trainer.seed
-    # Mirrors key splitting in train_lm.py: data_key, loader_key, model_key, training_key
-    data_key, _, _, _ = jrandom.split(jrandom.PRNGKey(seed), 4)
-    if train_config.data_seed is not None:
-        data_key = jrandom.PRNGKey(train_config.data_seed)
-
-    train_length = train_config.train_seq_len
-    if train_length is None:
-        train_length = train_config.model.max_seq_len
-    Pos = hax.Axis("position", train_length)
-
-    mix_key, shuffle_key = jrandom.split(data_key)
-    batch_schedule = train_config.trainer.batch_schedule
-    initial_batch_size = batch_schedule.batch_size_at_step(0)
-
-    datasets = train_config.data.train_sets(Pos, key=shuffle_key, initial_batch_size=initial_batch_size)
-
-    weights = train_config.data.train_weights
-    if isinstance(weights, list):
-        weights = rescale_mixture_schedule_for_batch_schedule(weights, batch_schedule)
-
-    dataset = MixtureDataset(
-        datasets=datasets,
-        weights=weights,
-        stop_strategy=train_config.data.stop_strategy,
-        key=mix_key,
-        block_size=train_config.data.mixture_block_size,
-    )
-    return dataset, batch_schedule
-
-
-def _get_source_names(dataset: MixtureDataset, indices: list[int]) -> list[str]:
-    """Map global indices to their source dataset names."""
-    names = []
-    for idx in indices:
-        block_id = idx // dataset.block_size
-        index_within_block = idx % dataset.block_size
-        block = dataset._get_block(block_id)
-        packed_id = block[index_within_block]
-        dataset_id = packed_id >> 16
-        names.append(dataset.dataset_index[dataset_id])
-    return names
-
-
-def _get_step_weights(dataset: MixtureDataset, batch_schedule, step: int) -> dict[str, float]:
-    """Get the mixture weights active at a given step."""
-    offset = batch_schedule.global_data_offset_by_step(step)
-    block_id = offset // dataset.block_size
-    stage_idx = dataset._get_stage_for_block(block_id)
-    return dataset.weight_stages[stage_idx][1]
-
-
-def _fetch_step_examples(dataset, batch_schedule, tokenizer, step: int) -> list[dict]:
-    """Fetch and decode all examples for a given step, with metadata."""
-    offset = batch_schedule.global_data_offset_by_step(step)
-    bs = batch_schedule.batch_size_at_step(step)
-    indices = list(range(offset, offset + bs))
-    examples = blocking_wait(dataset.get_batch(indices))
-    sources = _get_source_names(dataset, indices)
-    results = []
-    for i, (ex, src) in enumerate(zip(examples, sources, strict=True)):
-        tok_arr = ex.tokens.array if hasattr(ex.tokens, "array") else ex.tokens
-        tokens = tok_arr.tolist()
-        lw = ex.loss_weight.array if hasattr(ex.loss_weight, "array") else ex.loss_weight
-        pct_masked = float((lw == 0).sum()) / lw.size * 100
-        results.append(
-            {
-                "index": i,
-                "text": tokenizer.decode(tokens),
-                "source": src,
-                "num_tokens": len(tokens),
-                "pct_masked": round(pct_masked, 1),
-            }
-        )
-    return results
-
-
-def _build_summary(dataset, batch_schedule, examples: list[dict], step: int) -> dict:
-    """Build a summary dict for a step's batch, including outlier detection."""
-    weights = _get_step_weights(dataset, batch_schedule, step)
-    source_counts = dict(Counter(ex["source"] for ex in examples))
-    token_counts = [ex["num_tokens"] for ex in examples]
-    mask_pcts = [ex["pct_masked"] for ex in examples]
-
-    # Detect outliers
-    outliers = []
-    for ex in examples:
-        scores = _compute_outlier_scores(ex["text"])
-        reason = _is_outlier(scores)
-        if reason is not None:
-            outliers.append(
-                {
-                    "index": ex["index"],
-                    "source": ex["source"],
-                    "reason": reason,
-                    "scores": scores,
-                    "text_preview": ex["text"][:300],
-                }
-            )
-
-    return {
-        "step": step,
-        "batch_size": len(examples),
-        "weights": weights,
-        "source_counts": source_counts,
-        "token_stats": {
-            "min": min(token_counts),
-            "max": max(token_counts),
-            "mean": round(sum(token_counts) / len(token_counts), 1),
-        },
-        "mask_stats": {
-            "min_pct": min(mask_pcts),
-            "max_pct": max(mask_pcts),
-            "mean_pct": round(sum(mask_pcts) / len(mask_pcts), 1),
-        },
-        "outlier_count": len(outliers),
-        "outliers": outliers,
-    }
-
-
-def _parse_steps(steps_str: str) -> list[int]:
-    """Parse a steps string like '0,100,500' or '0:1000:100'."""
-    if ":" in steps_str:
-        parts = steps_str.split(":")
-        if len(parts) == 2:
-            return list(range(int(parts[0]), int(parts[1])))
-        elif len(parts) == 3:
-            return list(range(int(parts[0]), int(parts[1]), int(parts[2])))
-        else:
-            raise click.ClickException(f"Invalid step range: {steps_str}")
-    return [int(s.strip()) for s in steps_str.split(",")]
-
-
-def _run_tui(fetch_fn, start_step: int):
-    """Interactive TUI: left/right for docs, up/down for steps, g for goto.
-
-    Args:
-        fetch_fn: callable(step: int) -> list[dict], returns example dicts for a step.
-    """
-    import curses
-    import textwrap
-
-    step = start_step
-    doc_idx = 0
-    examples: list[dict] = []
-    scroll_offset = 0
-    status_msg = ""
-    show_stats = False
-    goto_mode = False
-    goto_buf = ""
-
-    def load_step(scr):
-        nonlocal examples, doc_idx, scroll_offset, status_msg
-        h, w = scr.getmaxyx()
-        scr.clear()
-        scr.addnstr(h // 2, 0, f"Loading step {step}...", w - 1)
-        scr.refresh()
-        try:
-            examples = fetch_fn(step)
-            doc_idx = min(doc_idx, max(0, len(examples) - 1))
-            scroll_offset = 0
-            status_msg = ""
-        except Exception as e:
-            examples = []
-            status_msg = str(e)
-
-    def _source_distribution():
-        counts = Counter(ex["source"] for ex in examples)
-        total = len(examples)
-        parts = []
-        for name, count in sorted(counts.items(), key=lambda x: -x[1]):
-            pct = count / total * 100
-            parts.append(f"{name}: {count} ({pct:.1f}%)")
-        return " | ".join(parts)
-
-    def draw(scr):
-        scr.clear()
-        h, w = scr.getmaxyx()
-        if h < 4 or w < 20:
-            return
-
-        # Header with metadata
-        if examples and 0 <= doc_idx < len(examples):
-            ex = examples[doc_idx]
-            header = (
-                f" Step {step} | Doc {doc_idx + 1}/{len(examples)}"
-                f" [{ex['source']}] {ex['num_tokens']}tok {ex['pct_masked']}%masked"
-                f" | q:quit </>:docs ^/v:steps g:goto s:stats"
-            )
-        else:
-            header = f" Step {step} | Doc {doc_idx + 1}/{len(examples)} | q:quit </>:docs ^/v:steps g:goto"
-        scr.attron(curses.A_REVERSE)
-        scr.addnstr(0, 0, header.ljust(w), w - 1)
-        scr.attroff(curses.A_REVERSE)
-
-        if status_msg:
-            scr.addnstr(2, 0, f"Error: {status_msg}", w - 1)
-            scr.refresh()
-            return
-
-        if not examples:
-            scr.addnstr(2, 0, "No examples.", w - 1)
-            scr.refresh()
-            return
-
-        # Determine footer area height
-        footer_lines = 1
-        stats_line = ""
-        if show_stats:
-            stats_line = _source_distribution()
-            footer_lines = 2
-
-        # Wrap and display text with scrolling
-        text = examples[doc_idx]["text"]
-        wrapped = []
-        for line in text.split("\n"):
-            if line == "":
-                wrapped.append("")
-            else:
-                wrapped.extend(textwrap.wrap(line, w - 1) or [""])
-
-        body_h = h - 1 - footer_lines
-        visible = wrapped[scroll_offset : scroll_offset + body_h]
-        for row, line in enumerate(visible):
-            scr.addnstr(row + 1, 0, line, w - 1)
-
-        # Stats bar
-        if stats_line:
-            scr.attron(curses.A_REVERSE)
-            scr.addnstr(h - 2, 0, (" " + stats_line).ljust(w), w - 1)
-            scr.attroff(curses.A_REVERSE)
-
-        # Footer / goto prompt
-        if goto_mode:
-            footer = f" Go to step: {goto_buf}_ (Enter to confirm, Esc to cancel)"
-        else:
-            total_lines = len(wrapped)
-            if total_lines > body_h:
-                pct = min(100, int((scroll_offset + body_h) / total_lines * 100))
-                footer = f" {pct}%  PgUp/PgDn to scroll "
-            else:
-                footer = ""
-        scr.attron(curses.A_REVERSE)
-        scr.addnstr(h - 1, 0, footer.ljust(w), w - 1)
-        scr.attroff(curses.A_REVERSE)
-
-        scr.refresh()
-
-    def main_loop(scr):
-        nonlocal step, doc_idx, scroll_offset, show_stats, goto_mode, goto_buf
-        curses.curs_set(0)
-        scr.timeout(-1)
-        load_step(scr)
-        draw(scr)
-
-        while True:
-            key = scr.getch()
-
-            if goto_mode:
-                if key == 27:  # Escape
-                    goto_mode = False
-                    goto_buf = ""
-                elif key in (curses.KEY_ENTER, 10, 13):
-                    if goto_buf:
-                        try:
-                            step = int(goto_buf)
-                        except ValueError:
-                            pass
-                        load_step(scr)
-                    goto_mode = False
-                    goto_buf = ""
-                elif key == curses.KEY_BACKSPACE or key == 127:
-                    goto_buf = goto_buf[:-1]
-                elif 48 <= key <= 57:  # digits 0-9
-                    goto_buf += chr(key)
-                draw(scr)
-                continue
-
-            if key == ord("q") or key == ord("Q"):
-                break
-            elif key == ord("g") or key == ord("G"):
-                goto_mode = True
-                goto_buf = ""
-            elif key == ord("s") or key == ord("S"):
-                show_stats = not show_stats
-            elif key == curses.KEY_RIGHT:
-                if examples and doc_idx < len(examples) - 1:
-                    doc_idx += 1
-                    scroll_offset = 0
-            elif key == curses.KEY_LEFT:
-                if doc_idx > 0:
-                    doc_idx -= 1
-                    scroll_offset = 0
-            elif key == curses.KEY_DOWN:
-                step += 1
-                load_step(scr)
-            elif key == curses.KEY_UP:
-                if step > 0:
-                    step -= 1
-                    load_step(scr)
-            elif key == curses.KEY_NPAGE:  # Page Down
-                h, w = scr.getmaxyx()
-                body_h = h - 1 - (2 if show_stats else 1)
-                if examples:
-                    text = examples[doc_idx]["text"]
-                    total = 0
-                    for ln in text.split("\n"):
-                        total += len(textwrap.wrap(ln, w - 1)) if ln else 1
-                    scroll_offset = min(scroll_offset + body_h, max(0, total - body_h))
-            elif key == curses.KEY_PPAGE:  # Page Up
-                h = scr.getmaxyx()[0]
-                body_h = h - 1 - (2 if show_stats else 1)
-                scroll_offset = max(0, scroll_offset - body_h)
-            elif key == curses.KEY_RESIZE:
-                pass
-            else:
-                continue
-            draw(scr)
-
-    curses.wrapper(main_loop)
-
-
-def _normalize_cluster_region(cluster: str) -> str:
-    """Extract the canonical region string from a --cluster argument.
-
-    Accepts region names like ``us-central2`` or ``marin-us-central2`` and
-    strips the ``marin-`` prefix if present.
-    """
-    region = cluster
-    if region.startswith("marin-"):
-        region = region[len("marin-") :]
-    return region
-
-
-def _validate_data_region(wandb_dict: dict, cluster: str) -> None:
-    """Ensure every component cache_dir lives in the same region as *cluster*.
-
-    Raises ``click.ClickException`` if any component's GCS bucket is in a
-    different region, preventing accidental cross-region egress charges.
-    """
-    from rigging.filesystem import REGION_TO_DATA_BUCKET
-
-    cluster_region = _normalize_cluster_region(cluster)
-
-    # Build reverse mapping: bucket name -> region
-    bucket_to_region: dict[str, str] = {bucket: region for region, bucket in REGION_TO_DATA_BUCKET.items()}
-
-    mismatches: list[str] = []
-    for name, cache_dir in wandb_dict.get("components", {}).items():
-        if not cache_dir or not cache_dir.startswith("gs://"):
-            continue
-        # Extract bucket name from gs://bucket/path
-        bucket_name = cache_dir.split("/")[2]
-        data_region = bucket_to_region.get(bucket_name)
-        if data_region is None:
-            # Unknown bucket — can't verify, warn but allow
-            logger.warning(
-                "Cannot determine region for bucket %r (component %r); skipping region check.",
-                bucket_name,
-                name,
-            )
-            continue
-        if data_region != cluster_region:
-            mismatches.append(
-                f"  - component {name!r}: bucket {bucket_name!r} is in {data_region}, "
-                f"but cluster is in {cluster_region}"
-            )
-
-    if mismatches:
-        detail = "\n".join(mismatches)
-        raise click.ClickException(
-            f"Cross-region data access detected! The data lives in a different region "
-            f"than the target cluster ({cluster_region}).\n"
-            f"This would incur significant egress charges.\n\n"
-            f"{detail}\n\n"
-            f"Please use --cluster <region> matching your data's region."
-        )
-
-
-def _resolve_cluster_config(cluster: str) -> str:
-    from marin.cluster.config import find_config_by_region
-
-    if cluster.endswith(".yaml") or os.path.exists(cluster):
-        return cluster
-    return find_config_by_region(cluster)
-
-
-def _cluster_runtime_env() -> dict:
-    from fray.v1.cluster.ray.deps import build_python_path
-
-    return {
-        "working_dir": os.getcwd(),
-        "excludes": [".git", "docs/", "**/*.pack", "lib/levanter/docs"],
-        "env_vars": {"PYTHONPATH": ":".join(build_python_path())},
-    }
-
-
-def _run_tui_on_cluster(cluster: str, wandb_config_json: str, step: int):
-    """Run the TUI locally with data fetched from a Ray cluster actor."""
-    import ray
-
-    from fray.v1.cluster.ray import DashboardConfig, ray_dashboard
-
-    config = _resolve_cluster_config(cluster)
-    runtime_env = _cluster_runtime_env()
-
-    with ray_dashboard(DashboardConfig.from_cluster(config)) as conn:
-        cluster_name = next(iter(conn.port_mappings))
-        api_port = conn.port_mappings[cluster_name].api_port
-        ray.init(address=f"ray://localhost:{api_port}", runtime_env=runtime_env)
-
-        @ray.remote
-        class DataInspector:
-            def __init__(self, config_json):
-                wandb_dict = json.loads(config_json)
-                train_config, _ = _config_from_wandb_dict(wandb_dict)
-                self.tokenizer = train_config.data.the_tokenizer
-                self.dataset, self.batch_schedule = _build_dataset(train_config)
-
-            def fetch_step(self, step):
-                return _fetch_step_examples(self.dataset, self.batch_schedule, self.tokenizer, step)
-
-        click.echo("Creating data inspector on cluster...", err=True)
-        inspector = DataInspector.remote(wandb_config_json)
-        # Warm up: ensure actor is ready before entering curses
-        ray.get(inspector.fetch_step.remote(step))
-        click.echo("Ready.", err=True)
-
-        def fetch_fn(s):
-            return ray.get(inspector.fetch_step.remote(s))
-
-        _run_tui(fetch_fn, step)
-
-
-@click.command()
-@click.option("--wandb-run", default=None, help="WandB run path (entity/project/run_id) or URL.")
-@click.option(
-    "--wandb-config-json",
-    default=None,
-    hidden=True,
-    help="Serialized WandB config (internal, used for cluster passthrough).",
-)
-@click.option("--step", default=None, type=int, help="Training step to inspect.")
-@click.option("--steps", default=None, type=str, help="Comma-separated steps or start:end:stride range.")
-@click.option("--output", "-o", default=None, type=click.Path(), help="Output JSONL path (default: stdout).")
-@click.option("--cluster", default=None, help="Ray cluster name or config path. Submits as a Ray job when set.")
-@click.option("--tui", is_flag=True, default=False, help="Interactive viewer.")
-@click.option("--truncate", default=None, type=int, help="Max characters per text in output.")
-@click.option("--summary", is_flag=True, default=False, help="Print per-step stats instead of full text.")
-def main(
-    wandb_run: str | None,
-    wandb_config_json: str | None,
-    step: int | None,
-    steps: str | None,
-    output: str | None,
-    cluster: str | None,
-    tui: bool,
-    truncate: int | None,
-    summary: bool,
-):
-    """Dump decoded training examples for a given step's batch to JSONL."""
-    if wandb_run is None and wandb_config_json is None:
-        raise click.ClickException("Must specify --wandb-run.")
-    if wandb_run is not None and wandb_config_json is not None:
-        raise click.ClickException("Cannot use both --wandb-run and --wandb-config-json.")
-
-    if step is None and steps is None:
-        if tui:
-            step = 0
-        else:
-            raise click.ClickException("Must specify --step or --steps.")
-    if step is not None and steps is not None:
-        raise click.ClickException("Cannot use both --step and --steps.")
-    if tui and steps is not None:
-        raise click.ClickException("--steps cannot be used with --tui. Use --step instead.")
-
-    step_list = _parse_steps(steps) if steps else [step]
-
-    # When submitted as a Ray job, the script runs on the cluster without --cluster.
-    # Detect this via RAY_JOB_ID which Ray sets automatically for submitted jobs.
-    on_cluster_node = os.environ.get("RAY_JOB_ID") is not None or os.environ.get("MARIN_PREFIX") is not None
-
-    if not cluster and not on_cluster_node:
-        raise click.ClickException("Must specify --cluster.")
-
-    # Resolve --wandb-run locally (where we have credentials) and pass the
-    # serialized config to the cluster as --wandb-config-json.
-    if wandb_run and cluster:
-        click.echo(f"Fetching config from WandB run: {wandb_run}", err=True)
-        wandb_config_json = json.dumps(_fetch_wandb_config(wandb_run))
-        wandb_run = None
-
-    # Validate that the data's region matches the cluster before submitting,
-    # to avoid cross-region egress charges.
-    if cluster and wandb_config_json:
-        _validate_data_region(json.loads(wandb_config_json), cluster)
-
-    if cluster and tui:
-        _run_tui_on_cluster(cluster, wandb_config_json, step_list[0])
-        return
-
-    if cluster:
-        _submit_to_cluster(
-            cluster,
-            step,
-            output,
-            steps=steps,
-            truncate=truncate,
-            summary=summary,
-            wandb_config_json=wandb_config_json,
-        )
-        return
-
-    # Running on a cluster node as a submitted Ray job.
-    if wandb_config_json:
-        wandb_dict = json.loads(wandb_config_json)
-    else:
-        wandb_dict = _fetch_wandb_config(wandb_run)
-
-    train_config, run_id = _config_from_wandb_dict(wandb_dict)
-    marin_prefix = os.environ.get("MARIN_PREFIX", "")
-    step_output_path = os.path.join(marin_prefix, run_id) if marin_prefix else run_id
-
-    tokenizer = train_config.data.the_tokenizer
-
-    dataset, batch_schedule = _build_dataset(train_config)
-
-    out = open(output, "w") if output else sys.stdout
-    total_examples = 0
-    try:
-        for s in step_list:
-            examples = _fetch_step_examples(dataset, batch_schedule, tokenizer, s)
-            if summary:
-                json.dump(_build_summary(dataset, batch_schedule, examples, s), out)
-                out.write("\n")
-            else:
-                for ex in examples:
-                    record = dict(ex, step=s)
-                    if truncate is not None and len(record["text"]) > truncate:
-                        record["text"] = record["text"][:truncate]
-                        record["truncated"] = True
-                    json.dump(record, out)
-                    out.write("\n")
-            total_examples += len(examples)
-    finally:
-        if output:
-            out.close()
-
-    if output and on_cluster_node:
-        import gcsfs
-
-        gcs_output = os.path.join(step_output_path, "debug", output)
-        fs = gcsfs.GCSFileSystem()
-        fs.put(output, gcs_output)
-        click.echo(f"Wrote {total_examples} examples to {gcs_output}")
-        click.echo(f"Download with: gsutil cp {gcs_output} .")
-    elif output:
-        click.echo(f"Wrote {total_examples} examples to {output}")
-
-
-if __name__ == "__main__":
-    main()
